1.update new route structure
